Fix Debug/Verbose preferences, fix Invoke-RestMethod on PS5
The Invoke-RestMethod parameter -HttpVersion was added in PowerShell 7.3
so that we can use HTTP/3 for DoH calls, as the specification requires
at least HTTP/2.  However, older versions of PowerShell can only use
HTTP/1.1.  I've refactored that code so that HTTP/3 is used when it is
available, and on older platforms, it gracefully falls back to HTTP/1.1
instead of failing with a parameter not found error.

Additionally, I realized -Debug and -Verbose preferences may not be
trickling down into other function calls.  Now, DebugPreference is
explicitly passed into Invoke-GooglePublicDnsApi, and VerbosePreference
is passed when supported.  In Test-MailPolicy, script-level Debug and
Verbose preferences are set at the script level.