Add RSA key length guessing for .NET <7
Turns out the ImportFromPem() method was added in .NET 7.  This is
required to accurately detect the number of bits in an RSA key.

As an alternative, I sat around and generated some RSA keys, then
converted them into Base64.  2048-bit keys have 392 characters,
1024-bit keys have 216 characters, 768-bit keys have 168 characters,
and 512-bit keys have 128 characters.  On Windows PowerShell or
PowerShell Core 6, we will use these guesses to determine the size
of an RSA DKIM key.

If anyone has a better way to guess that works on all platforms,
please submit a pull request.