Check that servers have both IPv4 and IPV6 addresses.
This commit adds a new feature that I'd considered during development.
Test-MailPolicy now calls Test-IPVersions, which will make sure that a
given hostname has both A and AAAA resource records in DNS.

This test ensures that a given server is reachable by clients that may
have only an IPv4 address (i.e., behind an old router) or only an IPv6
address (i.e., on a NAT64 network using an IPv4-only app). Note that we
do not attempt to connect to the server to verify that it is actually
reachable, as the PowerShell host may not be running on a computer with
both address families, or IPv6 may be broken.