Fix and enhance SPF redirect= test.
In versions 1.1.0 and older, the SPF tester would fail to parse the
"redirect" modifier, instead showing a null value on the screen.

This commit not only fixes that, but replaces that entirely with a
recursive test that tests the value of the "redirect" modifier.

This does introduce a possible denial-of-service issue where the
app could enter into an infinite loop due to a misconfiguration (or
intentional configuration) such as this:

domainA.com. IN TXT "v=spf1 redirect=domainB.com"
domainB.com. IN TXT "v=spf1 redirect=domainA.com"

The cmdlet should have a way to keep track of DNS lookups and return
a PermError after the tenth.  A future version of this will keep
track of how many DNS lookups are performed across all of the tokens
and modifiers specified (i.e., `redirect`, `a`, `mx`, `include`).