Add support for recursive SPF evaluation.
This commit adds support for running the SPF evaluations recursively,
to make sure that no more than ten additional DNS lookups will need
to be performed.  At that point, per the RFC, parsers may choose to
abort and return a PermError.  This new behavior can be invoked with
Test-SpfRecord -CountDnsLookups or Test-MailPolicy -CountSpfDnsLookups.

For consistency, this does revert the new feature of version 1.2.0;
redirect= modifiers will not be evaluated unless we are running in
recursive mode.

As a result of having more eyes on my own code, many small typos and
grammatical errors in the SPF output have been corrected.

It is possible that a future major version of MailPolicyExplainer
will make recursive mode the default.

Update CHANGELOG to note new recursion behavior.

The CHANGELOG has been updated to show that the commit to add recursive
SPF evaluation breaks the new behavior introduced in version 1.2.0.

Signed-off-by: Colin Cogle <colin@colincogle.name>