Fix DANE parser's handling of domains without MX records.
This commit fixes `Test-DaneRecords` to properly support domains that,
confusingly, do not specify MX records. In this case (which is spelled
out in the RFC), the A and AAAA records of the domain itself (`@`) are
used to construct an implied MX host with preference 0.

In previous versions, a DANE check would mistakenly be performed on a
null domain (`""`) instead of the domain name itself.