Check for TLSRPT "v" tag, improve error handling.
The SMTP TLS Reporting test now checks to make sure that there is
exactly one "v" tag and that it has the value of "TLSRPTv1". If not,
an error is returned, as the record is not valid per the RFC.

In addition, I came across a domain with a wildcard DNS record, so
their (nonexistent) TLSRPT policy was random text.  That showed me
the parser failed whenever it did not find a "token" with an equals
sign in it.  This has also been corrected.