Restore PS 5.1 compatibility.
Due to a missing backtick, this module was no longer loading on downlevel
versions of PowerShell (specifically, Windows PowerShell 5.1).  The older
versions considered this a syntax error, despite PowerShell 7 loading and
parsing the .psm1 file correctly.

This commit fixes that by adding one missing backtick in the right place.

Thank you to my colleague, Aslan Grealis, for finding this flaw.