Remove ip4: qualifier from output.
Due to a parser bug, IPv4 addresses extracted from the `ip4:` tokens
would have the qualifier prepended to the address.  This has been
corrected by fixing the parser to exclude any given qualifier.

This was never an issue for parsing IPv6 addresses from the `ip6:`
tokens, because I managed to write that code correctly.  #legacyIP