Add easy parameter to check ExO DKIM selectors.
This commit was an idea I've been mulling over for a while.  There is
now a `Test-MailPolicy -ExchangeOnline` parameter. This is the same as
`Test-MailPolicy -DkimSelectorsToCheck selector1,selector2`. Of course,
this doesn't prevent you from also using `-DkimSelectorsToCheck` to
check additional selectors at the same time.

Shout out to the thousands of Exchange Online administrators and support
staff, including the dozens that I work with.  You're welcome.

Closes: #6

Format: markdown
Issue: 6
Reviewer: github
Signed-off-by: Colin Cogle <colin@colincogle.name>
Tested-by: Colin Cogle <colin@colincogle.name>